// modules/payment-suppliers/managment-payments/enums/ps-table-keys.enum.ts
export enum SPTableKeys {
  COLUMN = 'column',
  FIELD = 'field',
  EXAMPLE = 'example',
  LIMIT = 'limit',
}

export enum SPTableKeys {
  ACCOUNT = 'cuentaDestino',
  NAME = 'nombreCuenta',
  STATUS = 'estadoCuenta',
  EMAIL = 'correo',
  DETAIL = 'detalle',
  AMOUNT = 'montoDestino',
}

export enum SPTableTrackKeys {
  ACCOUNT = 'cuenta',
  NAME = 'nombreCuenta',
  STATUS = 'statusCuenta',
  EMAIL = 'correo',
  DETAIL = 'detalle',
  AMOUNT = 'monto',
}